web release 1.3

+++ backend/src/main/java/org/kaleta/service/FirebaseService.java
@@ -27,55 +27,59 @@ public class FirebaseService
     private static FirebaseDatabase database;
 
     static {
-        String databaseUrl = ConfigProvider.getConfig().getValue("firebase.db.url", String.class);
+        if (ConfigProvider.getConfig().getValue("environment", String.class).equals("PRODUCTION"))
+        {
+            FirebaseOptions options;
+            try {
+                InputStream serviceKeyStream = FirebaseService.class.getResourceAsStream("/cert.json");
 
-        FirebaseOptions options;
-        try {
-            InputStream serviceKeyStream = FirebaseService.class.getResourceAsStream("/cert.json");
+                options = new FirebaseOptions.Builder()
+                        .setCredentials(GoogleCredentials.fromStream(serviceKeyStream))
+                        .setDatabaseUrl(ConfigProvider.getConfig().getValue("firebase.db.url", String.class))
+                        .build();
+            } catch (IOException e) {
+                throw new ServiceFailureException(e);
+            }
 
-            options = new FirebaseOptions.Builder()
-                    .setCredentials(GoogleCredentials.fromStream(serviceKeyStream))
-                    .setDatabaseUrl(databaseUrl)
-                    .build();
-        } catch (IOException e) {
-            throw new ServiceFailureException(e);
-        }
-
-        FirebaseApp app;
-        if (FirebaseApp.getApps().size() == 0){
-            app = FirebaseApp.initializeApp(options);
-        } else {
-            app = FirebaseApp.getInstance();
-        }
-        database = FirebaseDatabase.getInstance(app);
-
-        database.getReference("company").addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                companies.clear();
-                for (DataSnapshot postSnapshot : dataSnapshot.getChildren()) {
-                    try {
-                        FirebaseCompany company = postSnapshot.getValue(FirebaseCompany.class);
-                        companies.add(company);
-                    } catch (Exception e) {
-                        e.printStackTrace();
+            FirebaseApp app;
+            if (FirebaseApp.getApps().size() == 0){
+                app = FirebaseApp.initializeApp(options);
+            } else {
+                app = FirebaseApp.getInstance();
+            }
+            database = FirebaseDatabase.getInstance(app);
+            database.getReference("company").addValueEventListener(new ValueEventListener() {
+                @Override
+                public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                    companies.clear();
+                    for (DataSnapshot postSnapshot : dataSnapshot.getChildren()) {
+                        try {
+                            FirebaseCompany company = postSnapshot.getValue(FirebaseCompany.class);
+                            companies.add(company);
+                        } catch (Exception e) {
+                            e.printStackTrace();
+                        }
                     }
                 }
-            }
-            @Override
-            public void onCancelled(@NonNull DatabaseError databaseError) {}
-        });
+                @Override
+                public void onCancelled(@NonNull DatabaseError databaseError) {}
+            });
+        }
     }
 
     private static List<FirebaseCompany> companies = new ArrayList<>();
 
     public boolean hasCompany(String ticker)
     {
+        if (!ConfigProvider.getConfig().getValue("environment", String.class).equals("PRODUCTION")) return false;
+
         return companies.stream().filter(company -> company.getTicker().equals(ticker)).collect(Collectors.toSet()).size() > 0;
     }
 
     public FirebaseCompany getCompany(String ticker)
     {
+        if (!ConfigProvider.getConfig().getValue("environment", String.class).equals("PRODUCTION")) return null;
+
         return companies.stream()
                 .filter(company -> company.getTicker().equals(ticker))
                 .findFirst()
@@ -84,6 +88,8 @@ public class FirebaseService
 
     public void pushAssets(List<Trade> activeTrades)
     {
+        if (!ConfigProvider.getConfig().getValue("environment", String.class).equals("PRODUCTION")) return;
+
         DatabaseReference db = database.getReference("asset");
         db.removeValueAsync();
         for (Trade trade : activeTrades)
